LowCostStationProject/SensorStation/Server/requirements.txt
django
django-cors-headers
djangorestframework
orjson
unidecode
dash
pandas
